fix(desktop): read the newest pair-scoped harness log
Harnesses became per (agent, relay) pair and now write
`agents/logs/{pubkey}__{relay_hash}.log`, but `get_managed_agent_log`
still read the legacy `agents/logs/{pubkey}.log`. The harness-log panel
therefore froze at each agent's last single-runtime line while live logs
accumulated in files it never opened.

The reader now selects the most recently modified log belonging to the
agent — pair-scoped or legacy — so agents that have not restarted since
the change keep working and the panel follows the active harness. The
response already carries `log_path`, so the panel names the file shown.

`storage.rs`'s inline test module moves to a `#[path]`-included sibling,
matching `teams.rs` and `archive/mod.rs`, which drops both halves under
the desktop file-size limit.
